requirements.txt
Django==5.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
drf-yasg